Replace ushort*/uint* with char16_t*/char32_t* in private API [1]
Task-number: QTBUG-110403
Pick-to: 6.5

+++ src/corelib/io/qdebug.h
@@ -100,7 +100,7 @@ public:
     inline QDebug &operator<<(char t) { stream->ts << t; return maybeSpace(); }
     inline QDebug &operator<<(signed short t) { stream->ts << t; return maybeSpace(); }
     inline QDebug &operator<<(unsigned short t) { stream->ts << t; return maybeSpace(); }
-    inline QDebug &operator<<(char16_t t) { return *this << QChar(ushort(t)); }
+    inline QDebug &operator<<(char16_t t) { return *this << QChar(t); }
     inline QDebug &operator<<(char32_t t) { putUcs4(t); return maybeSpace(); }
     inline QDebug &operator<<(signed int t) { stream->ts << t; return maybeSpace(); }
     inline QDebug &operator<<(unsigned int t) { stream->ts << t; return maybeSpace(); }
